add policies, secure category module

+++ app/Domains/Catalog/Application/Actions/Category/DeleteCategoryAction.php
@@ -5,18 +5,17 @@ namespace App\Domains\Catalog\Application\Actions\Category;
 use App\Domains\Catalog\Domain\Exceptions\CategoryHasProductsException;
 use App\Domains\Catalog\Infrastructure\Models\Category;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Validation\ValidationException;
 
 class DeleteCategoryAction
 {
     public function execute(Category $category): ?bool
     {
+        Gate::authorize('delete', $category);
         if ($category->products()->exists()) {
             throw new CategoryHasProductsException();
         }
-
-        return DB::transaction(function () use ($category) {
-            return $category->delete();
-        });
+        return $category->delete();
     }
 }
