+++ routes/api.php
@@ -3,10 +3,14 @@
 use App\Domains\Catalog\Presentation\Http\Controllers\CategoryController;
 use Illuminate\Support\Facades\Route;
 
-Route::prefix('admin/v1')->group(function () {
-    Route::get('categories', [CategoryController::class, 'index'])->name('categories.index');
-    Route::post('categories', [CategoryController::class, 'store'])->name('categories.store');
-    Route::get('categories/{category}', [CategoryController::class, 'show'])->name('categories.show');
-    Route::put('categories/{category}', [CategoryController::class, 'update'])->name('categories.update');
-    Route::delete('categories/{category}', [CategoryController::class, 'destroy'])->name('categories.destroy');
+
+Route::prefix('v1')->group(function () {
+    Route::get('/categories', [CategoryController::class, 'index'])->name('categories.index');
+    Route::get('/categories/{category}', [CategoryController::class, 'show'])->name('categories.show');
+});
+
+Route::prefix('v1/admin')->middleware(['admin'])->group(function () {
+    Route::post('/categories', [CategoryController::class, 'store'])->name('admin.categories.store');
+    Route::put('/categories/{category}', [CategoryController::class, 'update'])->name('admin.categories.update');
+    Route::delete('/categories/{category}', [CategoryController::class, 'destroy'])->name('admin.categories.destroy');
 });
